call alert handler for each alert
do nothing for alert if there is no associated handler

+++ samples/r_alert_handler.R
@@ -2,48 +2,58 @@
 #   R script to generate an email report for an alert.
 #
 #   set the following constants to send email.
-emailServer <- "emailTest.sl.com"
-recipients  <- c("someone@sl.com")  # list of recipients
-acctName    <- "sl\\someone"        # need a valid email account to send mail
-acctPassword <- "----------"        # password for email account
+emailServer <- "emailTest.company.com"
+recipients  <- c("someone@company.com")  # list of recipients
+acctName    <- "company\\someone"        # need a valid email account to send mail
+acctPassword <- "password"        # password for email account
 
-# set location of RTView REST server for data fetched by template
-rtvDataserver <- "localhost:8068"
+# set default location of RTView REST server for data fetched by alert handlers
+rtvDataserver <- "rtvdemos-163.sl.com"
+rtvquery <- "simdata_rtvquery"
+
 #######################################################################
+inputArgs <- commandArgs(trailingOnly = TRUE)
+#inputArgs <- c("CENTRAL","VmwVmCpuUtilizationHigh","vSphere2~VMIRIS1061",1035,1,"High Warning Limit exceeded current value: 69.26 limit: 50.0")
 
 source("sl_utils.R")
-inputArgs <- commandArgs(trailingOnly = TRUE)
-tod <- Sys.time()
+email_template <- paste("alert_handlers/",inputArgs[2],'.Rmd',sep="")
 
-library(knitr)
-# use knitr to create html report from a R-markdown template
-# make filenames unique to handle alerts concurrently
-mdFileName <- paste('sample_email_', Sys.getpid(), '.md', sep="")
-htmlFileName <- paste('sample_email_', Sys.getpid(), '.html', sep="")
-knit('sample_email.Rmd', output=mdFileName)
+# ignore alert if there is no handler for it.
+if(file_test("-f",email_template)) {
 
-library(markdown)
-markdownToHTML(mdFileName, output=htmlFileName, options="")
+	library(knitr)
+	# use knitr to create html report from a R-markdown template
+	# make filenames unique to handle alerts concurrently
+	# (WARNING: your CPU could become quite busy if there is an "alert storm")
+	mdFileName <- paste(inputArgs[2], '_', Sys.getpid(), '.md', sep="")
+	htmlFileName <- paste(inputArgs[2], '_', Sys.getpid(), '.html', sep="")
+	knit(email_template, output=mdFileName)
 
-#define function to send email
-Sys.setenv(JAVA_HOME="")
-library(mailR)
-sendMail <- function(sender, recipients, subject, message) {
-    send.mail(from = sender,
-              to = recipients,
-              subject = subject,
-              body = message,
-              html = TRUE, inline = TRUE,
-              smtp = list(host.name = emailServer, port = 25, 
-                          user.name=acctName, passwd=acctPassword),
-              authenticate = TRUE, #debug = TRUE,
-              send = TRUE)
+	library(markdown)
+	markdownToHTML(mdFileName, output=htmlFileName, options="")
+
+	#define function to send email
+	Sys.setenv(JAVA_HOME="")
+	library(mailR)
+	sendMail <- function(sender, recipients, subject, message) {
+		send.mail(from = sender,
+				  to = recipients,
+				  subject = subject,
+				  body = message,
+				  html = TRUE, inline = TRUE,
+				  smtp = list(host.name = emailServer, port = 25, 
+							  user.name=acctName, passwd=acctPassword),
+				  authenticate = TRUE, #debug = TRUE,
+				  send = TRUE)
+	}
+
+	message <- paste(readLines(htmlFileName), collapse="\n")
+
+	sendMail("RTView EM <noreply@sl.com>", recipients, 
+			 paste("SL RTView EM Alert: ",inputArgs[2],sep=""), message)
+
+	unlink(htmlFileName)   # delete html report
+	unlink(mdFileName)   # delete markdown
+} else {
+    print("no handler")
 }
-
-message <- paste(readLines(htmlFileName), collapse="\n")
-
-sendMail("RTView EM <noreply@sl.com>", recipients, 
-         paste("EM Alert: ",inputArgs[2],sep=""), message)
-
-unlink(htmlFileName)   # delete html report
-unlink(mdFileName)   # delete markdown
